requirements.txt
aiogram==3.3.0
SQLAlchemy==2.0.28
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.1
pytesseract==0.3.10
python-dotenv==1.0.0
pillow==10.2.0
loguru==0.7.2 